# src/swagger.yaml
openapi: 3.1.0
info:
  title: Swagger Mini_Projet
  description: Swagger de l'api Mini_Projet
  version: 1.0.0
servers:
  - url: 'http://localhost:3000'
    description: Serveur de développement
paths:
...
  /events:
    post:
      summary: Création d'un événement sportif
      description: Permet de créer un nouvel événement sportif en fournissant les informations nécessaires.
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - sport
                - date
                - location
              properties:
                name:
                  type: string
                  example: "Tournoi de football"
                sport:
                  type: string
                  example: "Football"
                date:
                  type: string
                  format: date-time
                  example: "2024-12-15T10:00:00Z"
                location:
                  type: string
                  example: "Stade Pierre-Mauroy"
                description:
                  type: string
                  example: "Un tournoi de football amical."
      responses:
        '201':
          description: Événement créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Événement créé avec succès."
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "uuid-1234-5678-9101"
                      name:
                        type: string
                        example: "Tournoi de football"
                      sport:
                        type: string
                        example: "Football"
                      date:
                        type: string
                        format: date-time
                        example: "2024-12-15T10:00:00Z"
                      location:
                        type: string
                        example: "Stade Pierre-Mauroy"
        '400':
          description: Erreur de validation des données
        '500':
          description: Erreur interne du serveur

    get:
      summary: Liste des événements sportifs
      description: Permet de récupérer la liste de tous les événements sportifs.
      operationId: getAllEvents
      responses:
        '200':
          description: Liste des événements
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "uuid-1234-5678-9101"
                    name:
                      type: string
                      example: "Tournoi de football"
                    sport:
                      type: string
                      example: "Football"
                    date:
                      type: string
                      format: date-time
                      example: "2024-12-15T10:00:00Z"
                    location:
                      type: string
                      example: "Stade Pierre-Mauroy"
        '500':
          description: Erreur lors de la récupération des événements
  /events/{id}:
    get:
      summary: Détail d'un événement
      description: Permet de récupérer les détails d'un événement spécifique en fonction de son ID.
      operationId: getEventById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: L'ID de l'événement à récupérer
      responses:
        '200':
          description: Détails de l'événement
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "uuid-1234-5678-9101"
                  name:
                    type: string
                    example: "Tournoi de football"
                  sport:
                    type: string
                    example: "Football"
                  date:
                    type: string
                    format: date-time
                    example: "2024-12-15T10:00:00Z"
                  location:
                    type: string
                    example: "Stade Pierre-Mauroy"
        '404':
          description: Événement non trouvé
        '500':
          description: Erreur lors de la récupération de l'événement
    put:
      summary: Mise à jour d'un événement sportif
      description: Permet de mettre à jour un événement existant en fournissant les nouvelles informations.
      operationId: updateEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: L'ID de l'événement à mettre à jour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Tournoi de football"
                sport:
                  type: string
                  example: "Football"
                date:
                  type: string
                  format: date-time
                  example: "2024-12-15T10:00:00Z"
                location:
                  type: string
                  example: "Stade Pierre-Mauroy"
                description:
                  type: string
                  example: "Un tournoi de football amical."
      responses:
        '200':
          description: Événement mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Événement mis à jour avec succès."
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "uuid-1234-5678-9101"
        '404':
          description: Événement non trouvé
        '500':
          description: Erreur lors de la mise à jour de l'événement
    delete:
      summary: Suppression d'un événement sportif
      description: Permet de supprimer un événement sportif en fonction de son ID.
      operationId: deleteEvent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: L'ID de l'événement à supprimer
      responses:
        '204':
          description: Événement supprimé avec succès
        '404':
          description: Événement non trouvé
        '500':
          description: Erreur lors de la suppression de l'événement
  /events/register:
    post:
      summary: Inscription d'un participant à un événement
      description: Permet d'inscrire un participant à un événement spécifique.
      operationId: registerParticipant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
                - userId
              properties:
                eventId:
                  type: string
                  example: "uuid-1234-5678-9101"
                userId:
                  type: string
                  example: "uuid-5678-9101-1234"
      responses:
        '201':
          description: Participant inscrit avec succès
        '400':
          description: Champs manquants ou invalides
        '500':
          description: Erreur lors de l'inscription du participant
  /events/{eventId}/participants:
    get:
      summary: Liste des participants d'un événement
      description: Permet de récupérer la liste des participants d'un événement spécifique.
      operationId: getEventParticipants
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
          description: L'ID de l'événement
      responses:
        '200':
          description: Liste des participants de l'événement
        '500':
          description: Erreur lors de la récupération des participants
  /events/participants/{id}/cancel:
    post:
      summary: Annulation d'une inscription à un événement
      description: Permet d'annuler l'inscription d'un participant à un événement.
      operationId: cancelRegistration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: L'ID de l'inscription à annuler
      responses:
        '200':
          description: Inscription annulée avec succès
        '404':
          description: Inscription non trouvée
        '500':
          description: Erreur lors de l'annulation de l'inscription
...